Added __init__.py to webscraper, replaced webscraper.py with webscraper2.py, added placeholder date, add SQL insert
